Rename vehice_delegate to vehicle_delegate
