+++ shared/turbo_grid/widgets/vehicle_delegate.dart
